refactor: split out org views testing
Now that the code has been moved, we can move around the test code.
